AB-311: Use a single base test class for all tests
