+++ webserver/test/test_utils.py
@@ -1,8 +1,8 @@
-from webserver.testing import ServerTestCase
+from webserver.testing import AcousticbrainzTestCase
 from webserver import utils
 
 
-class UtilsTestCase(ServerTestCase):
+class UtilsTestCase(AcousticbrainzTestCase):
 
     def test_generate_string(self):
         length = 42
