+++ webserver/views/test/test_login.py
@@ -1,9 +1,9 @@
 from __future__ import absolute_import
-from webserver.testing import ServerTestCase
+from webserver.testing import AcousticbrainzTestCase
 from flask import url_for
 
 
-class LoginViewsTestCase(ServerTestCase):
+class LoginViewsTestCase(AcousticbrainzTestCase):
 
     def test_login_page(self):
         response = self.client.get(url_for('login.index'))
